Add files via upload
- Changed start text
- Changed order in menu
- [123] changed to [id123]
- Remove nickname
- Changed profile and order view
- Changed text in edit_profile
- Now you can press "Хочу стать испольнителем" button only one time
- Added new buttons in message like "User хочет стать испольнителем"
- Now you can only deposit more than 0руб
- Added deposit alert
- Fixed balance bags
- Added text for empty transactions list
- Changed order lists: subject, type, deadline, price
- Now bot sending only one message than order is finished
- New order view
- Changed text in "choose worker" message
- Then order is finished, chat is not active
- Now you can cancel your order only before status "Оплачен"
- Alert to worker than order is finished
- Changed all order statuses

+++ chat_panel.py
@@ -1,6 +1,6 @@
 import telegram
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup, ReplyKeyboardMarkup, InputMediaPhoto, \
-    InputMediaDocument
+    InputMediaDocument, ChatPermissions
 
 from menu import Menu
 from models import *
@@ -30,16 +30,32 @@ def chat(update, context):
 
                         description = 'Чат по поводу заказа ' + order.subject + ' [' + str(order.id) + ']'
                         context.bot.set_chat_description(chat_id=update.effective_chat.id, description=description)
+                        chat_permissions = ChatPermissions(
+                            can_send_messages=True,
+                            can_send_media_messages=True,
+                            can_send_polls=True,
+                            can_send_other_messages=True,
+                            can_add_web_page_previews=True,
+                            can_change_info=False,
+                            can_invite_users=False,
+                            can_pin_messages=False
+                        )
+                        context.bot.set_chat_permissions(chat_id=update.effective_chat.id, permissions=chat_permissions)
 
                         ftext = 'Установленно имя чата - ' + title + '\nУстановленно опсиание чата - ' + description
 
                         context.bot.send_message(chat_id=update.effective_chat.id, text=ftext, timeout=500)
 
                         link = context.bot.export_chat_invite_link(chat_id=update.effective_chat.id)
-                        context.bot.send_message(chat_id=int(order.user_id), text='Чат с исполнителем по поводу заказа ' +
+
+                        client = User.get(id=int(order.user_id))
+
+                        context.bot.send_message(chat_id=client.user_id, text='Чат с исполнителем по поводу заказа ' +
                                                                                   order.subject + ' [' + str(order.id)
                                                                                   + '] - ' + link)
-                        context.bot.send_message(chat_id=int(order.worker_id), text='Чат с заказчиком по поводу заказа '
+
+                        worker = User.get(id=int(order.worker_id))
+                        context.bot.send_message(chat_id=worker.user_id, text='Чат с заказчиком по поводу заказа '
                                                                                     + order.subject + ' [' + str(order.id)
                                                                                     + '] - ' + link)
 
@@ -65,15 +81,16 @@ def admin(update, context):
         user = get_user(update.message.from_user.id)
         if user:
             chat = Chat.get(chat_id=update.effective_chat.id)
-            order = Order.get(id=int(chat.order_id))
-            link = context.bot.export_chat_invite_link(chat_id=update.effective_chat.id)
-            text = 'Чат ' + str(chat.id) + ' вызывает админа по поводу заказа ' + order.subject + ' [' + str(order.id) + ']'
-            text += '\n' + link
+            if chat:
+                order = Order.get(id=int(chat.order_id))
+                link = context.bot.export_chat_invite_link(chat_id=update.effective_chat.id)
+                text = 'Чат ' + str(chat.id) + ' вызывает админа по поводу заказа ' + order.subject + ' [' + str(order.id) + ']'
+                text += '\n' + link
 
-            admins = list(select(u for u in User if u.status == 'admin'))
+                admins = list(select(u for u in User if u.status == 'admin'))
 
-            for a in admins:
-                context.bot.send_message(chat_id=a.id, text=text)
+                for a in admins:
+                    context.bot.send_message(chat_id=a.user_id, text=text)
 
 
 @db_session
@@ -82,41 +99,42 @@ def price(update, context):
         user = get_user(update.message.from_user.id)
         if user:
             chat = Chat.get(chat_id=update.effective_chat.id)
-            order = Order.get(id=int(chat.order_id))
+            if chat:
+                order = Order.get(id=int(chat.order_id))
 
-            if context.args:
-                try:
-                    price = int(context.args[0])
-                    if chat.price == '0':
-                        if user:
-                            name = get_name(user)
+                if context.args:
+                    try:
+                        price = int(context.args[0])
+                        if chat.price == '0':
+                            if user:
+                                name = get_name(user)
 
-                        chat.worker_yes = 0
-                        chat.user_yes = 0
+                            chat.worker_yes = 0
+                            chat.user_yes = 0
 
-                        if chat.price_msg:
-                            context.bot.delete_message(chat_id=chat.chat_id, message_id=chat.price_msg)
+                            if chat.price_msg:
+                                context.bot.delete_message(chat_id=chat.chat_id, message_id=chat.price_msg)
 
-                        text = "Пользователь " + name + ' предлагает цену - <b>' + context.args[0] + 'р.</b>\nНажмите кнопку ниже, если вы согласны.'
-                        text += '\nСогласны:\n'
+                            text = "Пользователь " + name + ' предлагает цену - <b>' + context.args[0] + 'р.</b>\nНажмите кнопку ниже, если вы согласны.'
+                            text += '\nСогласны:\n'
 
-                        mymenu = Menu()
-                        buttons = [InlineKeyboardButton('Согласен', callback_data='@' + str(chat.id) + '@yes@' + context.args[0])]
+                            mymenu = Menu()
+                            buttons = [InlineKeyboardButton('Согласен', callback_data='@' + str(chat.id) + '@yes@' + context.args[0])]
 
-                        markup = mymenu.build_menu(buttons=buttons, n_cols=1, header_buttons=None, footer_buttons=None)
-                        reply_markup = InlineKeyboardMarkup(markup)
+                            markup = mymenu.build_menu(buttons=buttons, n_cols=1, header_buttons=None, footer_buttons=None)
+                            reply_markup = InlineKeyboardMarkup(markup)
 
-                        chat.price_msg = context.bot.send_message(chat_id=update.effective_chat.id, text=text, parse_mode=telegram.ParseMode.HTML, reply_markup=reply_markup).message_id
-                    else:
-                        text = 'Цена для этого заказа уже утверждена!'
+                            chat.price_msg = context.bot.send_message(chat_id=update.effective_chat.id, text=text, parse_mode=telegram.ParseMode.HTML, reply_markup=reply_markup).message_id
+                        else:
+                            text = 'Цена для этого заказа уже утверждена!'
+                            context.bot.send_message(chat_id=update.effective_chat.id, text=text)
+                    except Exception as e:
+                        print(e)
+                        text = 'Используйте /price [price], [price] - целое число!'
                         context.bot.send_message(chat_id=update.effective_chat.id, text=text)
-                except Exception as e:
-                    print(e)
+                else:
                     text = 'Используйте /price [price], [price] - целое число!'
                     context.bot.send_message(chat_id=update.effective_chat.id, text=text)
-            else:
-                text = 'Используйте /price [price], [price] - целое число!'
-                context.bot.send_message(chat_id=update.effective_chat.id, text=text)
 
 
 @db_session
@@ -125,32 +143,33 @@ def done(update, context):
         user = get_user(update.message.from_user.id)
         if user:
             chat = Chat.get(chat_id=update.effective_chat.id)
-            order = Order.get(id=int(chat.order_id))
-            name = get_name(user)
+            if chat:
+                order = Order.get(id=int(chat.order_id))
+                name = get_name(user)
 
-            if order.status == 'Оплачен':
-                print(1)
-                if chat.worker_yes != 2 or chat.user_yes != 2:
-                    chat.worker_yes = 1
-                    chat.user_yes = 1
+                if order.status == 'Оплачен':
+                    print(1)
+                    if chat.worker_yes != 2 or chat.user_yes != 2:
+                        chat.worker_yes = 1
+                        chat.user_yes = 1
 
-                    if chat.done_msg:
-                        context.bot.delete_message(chat_id=chat.chat_id, message_id=chat.done_msg)
+                        if chat.done_msg:
+                            context.bot.delete_message(chat_id=chat.chat_id, message_id=chat.done_msg)
 
-                    text = "Пользователь " + name + ' предлагает завершить заказ.'
-                    text += '\nСогласны:\n'
+                        text = "Пользователь " + name + ' предлагает завершить заказ.'
+                        text += '\nСогласны:\n'
 
-                    mymenu = Menu()
-                    buttons = [InlineKeyboardButton('Согласен', callback_data='@' + str(chat.id) + '@done')]
+                        mymenu = Menu()
+                        buttons = [InlineKeyboardButton('Согласен', callback_data='@' + str(chat.id) + '@done')]
 
-                    markup = mymenu.build_menu(buttons=buttons, n_cols=1, header_buttons=None, footer_buttons=None)
-                    reply_markup = InlineKeyboardMarkup(markup)
+                        markup = mymenu.build_menu(buttons=buttons, n_cols=1, header_buttons=None, footer_buttons=None)
+                        reply_markup = InlineKeyboardMarkup(markup)
 
-                    chat.done_msg = context.bot.send_message(chat_id=update.effective_chat.id, text=text, parse_mode=telegram.ParseMode.HTML,
-                                             reply_markup=reply_markup).message_id
+                        chat.done_msg = context.bot.send_message(chat_id=update.effective_chat.id, text=text, parse_mode=telegram.ParseMode.HTML,
+                                                 reply_markup=reply_markup).message_id
+                    else:
+                        text = 'Заказ уже закрыт!'
+                        context.bot.send_message(chat_id=update.effective_chat.id, text=text)
                 else:
-                    text = 'Заказ уже закрыт!'
+                    text = 'Заказ еще не оплачен!'
                     context.bot.send_message(chat_id=update.effective_chat.id, text=text)
-            else:
-                text = 'Заказ еще не оплачен!'
-                context.bot.send_message(chat_id=update.effective_chat.id, text=text)
